update example, deprecate V1 open groups

+++ sample.js
@@ -27,34 +27,35 @@ client.loadIdentity({
     fs.writeFileSync('lastHash.txt', hash)
   })
 
-  //const openGroupName = 'session.lokisn.com'
-  //const openGroupName = 'feedback.getsession.org'
   const openGroupV2URL = 'http://open2.hesiod.network/build_a_bot?public_key=58dc124cc38e4d03449037e9a4a86a2e5c2a648938eb824a5cdf3b6a80fab07d'
-
-  //const ogOpened = await client.joinOpenGroup(openGroupName)
   const ogv2Handle = await client.joinOpenGroupV2(openGroupV2URL)
 
   // handle incoming messages
   client.on('messages', msgs => {
     msgs.forEach(async msg => {
-      /*
-      if (msg.openGroup) {
-        console.log(`New message, In group: ${msg.openGroup}`)
-      } else */
       if (msg.room) {
-        console.log(`New message, In group V2: ${msg.id} in ${msg.room}`)
+        console.log(`New message, In group: ${msg.id} in ${msg.room}`)
       } else {
         console.log('New message, Private')
       }
       console.log(`From: ${msg.profile && msg.profile.displayName} (${msg.source})`)
       console.log(msg.body)
 
+      // Download their avatar
+      // change 0 to 1 if you want to download avatars from users that you receive
+      if (0 && msg.profile && msg.profile.profilePicture) {
+        const avatarBuf = await client.decodeAvatar(msg.profile.profilePicture, msg.profileKey)
+        // write it to disk
+        fs.writeFileSync(msg.source + '.avatar', avatarBuf)
+      }
+
       // Attachment processing example
+      // change 0 to 1 if you want to download (the first) attachment sent to you
       if (0) {
         if (msg.attachments.length) {
           const attachments = await client.getAttachments(msg)
           //console.log('attachment', attachments[0])
-          fs.writeFileSync('imagetest.png', attachments[0])
+          fs.writeFileSync(msg.source + '.attachment', attachments[0])
         }
       }
       /*
@@ -81,6 +82,8 @@ client.loadIdentity({
   //
   // need an image
   //const attachment = await client.makeImageAttachment(fs.readFileSync('/Users/user2/Pictures/1587699732-0s.png'))
+  //
+  // change 0 to 1 if you want to send a message to SessionID
   if (0) {
     client.send(SessionID, 'Hello', {
       // attachments: [attachment]
@@ -88,20 +91,19 @@ client.loadIdentity({
       console.debug('Sent "Hello" to', SessionID)
     })
   }
-
-  /*
-  if (ogOpened) {
-    const messageId = await client.sendOpenGroupMessage(openGroupName, 'Sample Open Group Message 1')
-    await client.deleteOpenGroupMessage(openGroupName, [messageId])
+  // Open group invite example
+  // change 0 to 1 if you want to send an open group invite to SessionID
+  if (0) {
+    client.sendOpenGroupInvite(SessionID, 'Bob\'s server', '')
   }
-  */
 
+  // change 0 to 1 if you want to send a message to an open group
   if (0 && ogv2Handle) {
     const messageId = await client.sendOpenGroupV2Message(ogv2Handle, 'Hello World SOGSv2 from node-session-client')
-    await client.deleteOpenGroupV2Message(ogv2Handle, messageId)
+    // change 0 to 1 if you want to delete that message you sent to an open group
+    if (0) {
+      await client.deleteOpenGroupV2Message(ogv2Handle, messageId)
+    }
   }
 
-  // Open group invite example
-  // only works on desktop
-  //client.sendOpenGroupInvite(toPubkey, 'Bob\'s server', openGroupName, 2)
 })
